pyproject.toml
[project]
name = "dailywall"
version = "0.1.0"
description = "Bing Wallpaper Local Archive API"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.136.0",
    "uvicorn>=0.44.0",
    "sqlalchemy>=2.0.49",
    "pillow>=12.2.0",
    "httpx>=0.28.1",
    "pydantic-settings>=2.13.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0.3",
    "pytest-asyncio",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app", "crawler"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
